feat: role destroy cant delete when users are assigned to a role

+++ app/Http/Controllers/RoleController.php
@@ -133,6 +133,13 @@ class RoleController extends Controller
                 return $this->sendErrorOfNotFound404('Role not found');
             }
 
+            $role->withCount('users');
+            if ($role->users_count > 0) {
+                return $this->sendErrorOfUnprocessableEntity(
+                    "Cannot delete role because it is assigned to {$role->users_count} user(s)."
+                );
+            }
+
             $role->permissions()->detach();
             $role->delete();
